Add a third benchmark for warm start scenario
It appears that there are still things happening immediately after an
editor is opened that get in the way of benchmarking. So I created a
warm start scenario that shows while the new method is still faster, it
isn't faster by as great a margin as I first believed.

tabify-old-cold: 706.1650000000002
tabify-old-warm: 511.80999999999995
tabify-new: 477.99000000000024